Fix MFA enforcement redirect loop
